requirements.txt
PyPDF2
pdfplumber
python-docx